refactor(core): split skill-lab/index.ts into focused modules
Extract the 1554-line monolith into:
- schemas.ts: Zod validation schemas
- utils.ts: pure utility functions
- benchmark.ts: benchmark summarization and comparison logic
- skill-lab.ts: SkillLab class
- index.ts: thin re-export shim

Closes #43
